requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
pydantic==2.5.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic-settings==2.1.0
starlette-prometheus==0.9.0
prometheus-client==0.12.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
neo4j==5.14.1
redis==5.0.1
rq==1.15.1

# ML and NLP
#torch==2.1.0
#torchvision==0.16.0
transformers==4.35.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
rank-bm25==0.2.2

# Utilities
tqdm==4.66.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

#Testing Requirements
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0

schedule==1.2.0

PyMuPDF==1.19.6  # For PDF processing
langdetect==1.0.9  # For language detection
googletrans-py==4.0.0  # For translation
beautifulsoup4>=4.10.0  # For HTML processing

slowapi==0.1.9 # For rate-limiting
elasticsearch==9.0.1 # ELK
python-json-logger==3.3.0 # Logger
jwt==1.3.1
pika==1.3.2

# stress_testing/requirements.txt
requests>=2.28.1
pandas>=1.4.3
matplotlib>=3.5.2
numpy>=1.23.1
tqdm>=4.64.0
psycopg2-binary>=2.9.3
aiohttp>=3.8.1
locust>=2.11.0
pyjwt>=2.4.0